implementation/LiYancong/2019-07-23/Wit2Wit/requirements.txt
aliyun-python-sdk-core==2.13.5
aliyun-python-sdk-core-v3==2.13.3
aliyun-python-sdk-ecs==4.16.11
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.25
certifi==2019.6.16
chardet==3.0.4
Django==1.8.2
django-haystack==2.6.1
django-redis==4.4.0
django-tinymce==2.6.0
fdfs-client-py==1.2.6
idna==2.8
jieba==0.39
jmespath==0.9.4
kombu==3.0.37
mutagen==1.42.0
Pillow==6.1.0
pycryptodome==3.8.2
PyMySQL==0.9.3
pytz==2019.1
redis==2.10.6
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.3
vine==1.3.0
Whoosh==2.7.4
